server/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
distlib==0.3.4
Django==4.0.1
django-autoslug==1.9.8
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.2.0
filelock==3.4.2
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.1
PyJWT==2.3.0
pymongo==4.0.1
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
